Move tests to existing folders
