+++ app/components/Card/index.test.js
@@ -1,6 +1,6 @@
-import React from 'react';
-import { render, cleanup } from '@testing-library/react';
-import Card from '../Card';
+import React from "react";
+import { render, cleanup } from "@testing-library/react";
+import Card from "./index";
 
 describe("<Card />", () => {
   afterEach(cleanup);
